Move Ko test into a module of its own #54
